Added documentation tool
At the moment, things are hard-coded, but I will be moving away from
this in the near future, hopefully.

For some reason, I'm extremely pleased with how well DocTree turned out,
even though it wasn't that difficult of a task. Maybe I'm just happy I
was able to get a custom component to actually work in a reasonable
amount of time (I haven't had this much luck with android development,
to date).